# internal/bench: add benchmem flag support